+++ NEWS.md
@@ -5,6 +5,8 @@
 * `quick_dfl()` is now deprecated [#91](https://github.com/jmbarbone/fuj/issues/91)
 * `list()` is now a little faster [#91](https://github.com/jmbarbone/fuj/issues/91)
 * `list0()`'s functionality to ignore empty inputs can be disabled if `options(fuj.list.active = FALSE)` before `{fuj}` is loaded [#91](https://github.com/jmbarbone/fuj/issues/91)
+* `set_file_ext()` and `file_ext<-()` added for controlling file extensions [#89](https://github.com/jmbarbone/fuj/issues/89)
+* `+` and `/` methods added for `file_path` classes, allowing path creation (e.g., `fp("folder") / "subfolder" / "file" + "extension"`) [#89](https://github.com/jmbarbone/fuj/issues/89)
 * `vap` family functions added [#83](https://github.com/jmbarbone/fuj/issues/83)
   * `vaps` are vector apply functions, with certain presets to assist with common cases
   * all `vap` functions have type-stable variants: